web/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
whitenoise==4.1.4
Jinja2==2.10.1
dj-database-url==0.5.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1